+++ include/entities/entity_model.h
@@ -143,6 +143,7 @@ public:
   bool is_size_valid() const;
   virtual bool is_size_valid(const QSize& size) const;
   QSize get_closest_valid_size(const QSize& size) const;
+  QSize get_closest_snapped_size(const QSize& size) const;
   virtual QSize get_valid_size() const;
 
   // Displaying in the editor.